Fix MSVC 64bit warnings about int64 truncation in tests.

+++ tests/shared/filesystem.h
@@ -133,7 +133,7 @@ public:
 
         memset( reparseInfo, 0, sizeof( *reparseInfo ));
         reparseInfo->ReparseTag = IO_REPARSE_TAG_MOUNT_POINT;
-        reparseInfo->ReparseTargetLength = DWORD(target.size() * sizeof(wchar_t));
+        reparseInfo->ReparseTargetLength = WORD(target.size()) * WORD(sizeof(wchar_t));
         reparseInfo->ReparseTargetMaximumLength = reparseInfo->ReparseTargetLength + sizeof(wchar_t);
         target.toWCharArray(reparseInfo->ReparseTarget);
         reparseInfo->ReparseDataLength = reparseInfo->ReparseTargetLength + 12;
